leads-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.2
pydantic-settings==2.6.1
email-validator==2.2.0
python-multipart==0.0.12
minio==7.2.8
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.3.0
alembic==1.14.0
aiokafka==0.10.0